moved buttfilt back to ptsa.filt to avoid circular dependencies
