+++ ptsa/data/filters/ButterworthFilter.py
@@ -1,10 +1,12 @@
 from numpy import asarray
 from xarray import DataArray
+from ptsa.filt import buttfilt
 from ptsa.data.timeseries import TimeSeries
 from ptsa.data.common import get_axis_index
 from ptsa.data.filters import BaseFilter
 import traits.api
-from scipy.signal import butter, filtfilt
+
+__all__ = ['ButterworthFilter']
 
 class ButterworthFilter(BaseFilter):
     """Applies Butterworth filter to a time series.
@@ -61,36 +63,6 @@ class ButterworthFilter(BaseFilter):
         return filtered_time_series
 
 
-def buttfilt(dat,freq_range,sample_rate,filt_type,order,axis=-1):
-    """Wrapper for a Butterworth filter.
-
-    """
-
-    # make sure dat is an array
-    dat = asarray(dat)
-
-    # reshape the data to 2D with time on the 2nd dimension
-    #origshape = dat.shape
-    #dat = reshape_to_2d(dat,axis)
-
-    # set up the filter
-    freq_range = asarray(freq_range)
-
-    # Nyquist frequency
-    nyq=sample_rate/2.
-
-    # generate the butterworth filter coefficients
-    [b,a]=butter(order,freq_range/nyq,filt_type)
-
-    # loop over final dimension
-    #for i in range(dat.shape[0]):
-    #    dat[i] = filtfilt(b,a,dat[i])
-    #dat = filtfilt2(b,a,dat,axis=axis)
-    dat = filtfilt(b,a,dat,axis=axis)
-
-    # reshape the data back
-    #dat = reshape_from_2d(dat,axis,origshape)
-    return dat
 
 if __name__ == '__main__':
     import numpy as np
